Extract workspace thread APIs
