+++ Tests/QuillCodeParityTests/ParityWorkspaceSidebarGateTests.swift
@@ -3,6 +3,7 @@ import XCTest
 final class ParityWorkspaceSidebarGateTests: QuillCodeParityTestCase {
     func testWorkspaceModelDelegatesSidebarSelectionTransitions() throws {
         let modelText = try Self.appSourceText(named: "WorkspaceModel.swift")
+        let threadExtensionText = try Self.appSourceText(named: "WorkspaceModelThreads.swift")
         let selectionText = try Self.appSourceText(named: "WorkspaceSidebarSelectionEngine.swift")
         let bulkPlannerText = try Self.appSourceText(named: "WorkspaceSidebarBulkActionPlanner.swift")
         let bulkExecutorText = try Self.appSourceText(named: "WorkspaceSidebarBulkActionExecutor.swift")
@@ -13,17 +14,19 @@ final class ParityWorkspaceSidebarGateTests: QuillCodeParityTestCase {
         XCTAssertTrue(selectionText.contains("static func selectAll"), "Select-all behavior should be directly testable.")
         XCTAssertTrue(selectionText.contains("static func toggle"), "Selection toggles should be directly testable.")
         XCTAssertTrue(selectionText.contains("static func resolve"), "Stale-ID pruning and sidebar ordering should be directly testable.")
-        XCTAssertTrue(modelText.contains("WorkspaceSidebarSelectionEngine.start"), "WorkspaceModel should delegate selection start.")
-        XCTAssertTrue(modelText.contains("WorkspaceSidebarSelectionEngine.selectAll"), "WorkspaceModel should delegate select-all.")
-        XCTAssertTrue(modelText.contains("WorkspaceSidebarSelectionEngine.toggle"), "WorkspaceModel should delegate selection toggles.")
+        XCTAssertTrue(threadExtensionText.contains("WorkspaceSidebarSelectionEngine.start"), "WorkspaceModel thread/sidebar extension should delegate selection start.")
+        XCTAssertTrue(threadExtensionText.contains("WorkspaceSidebarSelectionEngine.selectAll"), "WorkspaceModel thread/sidebar extension should delegate select-all.")
+        XCTAssertTrue(threadExtensionText.contains("WorkspaceSidebarSelectionEngine.toggle"), "WorkspaceModel thread/sidebar extension should delegate selection toggles.")
         XCTAssertTrue(modelText.contains("WorkspaceSidebarSelectionEngine.resolve"), "WorkspaceModel should delegate stale-ID pruning and ordering.")
         XCTAssertTrue(bulkPlannerText.contains("struct WorkspaceSidebarBulkActionPlanner"), "Sidebar bulk action planning should live in a focused planner.")
         XCTAssertTrue(bulkPlannerText.contains("static func plan"), "Sidebar bulk action plans should be directly testable.")
         XCTAssertTrue(bulkPlannerText.contains("enum FollowUpSelection"), "Bulk action selection follow-up policy should be explicit.")
-        XCTAssertTrue(modelText.contains("WorkspaceSidebarBulkActionPlanner.plan"), "WorkspaceModel should delegate bulk action target planning.")
+        XCTAssertTrue(threadExtensionText.contains("WorkspaceSidebarBulkActionPlanner.plan"), "WorkspaceModel thread/sidebar extension should delegate bulk action target planning.")
         XCTAssertTrue(bulkExecutorText.contains("struct WorkspaceSidebarBulkActionExecutor"), "Sidebar bulk action execution should live in a focused executor.")
         XCTAssertTrue(bulkExecutorText.contains("static func execute"), "Sidebar bulk mutations should be directly testable.")
-        XCTAssertTrue(modelText.contains("WorkspaceSidebarBulkActionExecutor.execute"), "WorkspaceModel should delegate bulk action execution.")
+        XCTAssertTrue(threadExtensionText.contains("WorkspaceSidebarBulkActionExecutor.execute"), "WorkspaceModel thread/sidebar extension should delegate bulk action execution.")
+        XCTAssertFalse(modelText.contains("public func startSidebarSelection"), "WorkspaceModel.swift should not own sidebar-selection API bodies.")
+        XCTAssertFalse(modelText.contains("public func performSidebarBulkAction"), "WorkspaceModel.swift should not own sidebar bulk-action API bodies.")
         XCTAssertFalse(modelText.contains("public struct SidebarSelectionState"), "WorkspaceModel should not own sidebar selection state.")
         XCTAssertFalse(modelText.contains("selectedThreadIDs.insert"), "WorkspaceModel should not mutate sidebar selection sets directly.")
         XCTAssertFalse(modelText.contains("selectedThreadIDs.remove"), "WorkspaceModel should not mutate sidebar selection sets directly.")
